Todo/serializers.py
from rest_framework import serializers
from .models import Todo


class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = ['title', 'description', 'goal_set_date', 'set_to_complete', 'is_complete', 'user_id']