requirements.txt
anaconda==0.0.1.1
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
decorator==5.1.1
Django==5.1.3
django-allauth==65.4.1
django-rest-framework==0.1.0
django-split-settings==1.3.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-stubs==3.15.3
djangorestframework-yaml==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.16.1
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
Rest_Framework_Auth_Toolkit==0.13
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
